Moved some code from Worldmap.cpp to Interface.cpp
